OCEProject 3813 - Added MaximumRetries and ExposeRawResponse to make sure we can accurately capture errors for Search when any of the 3 nodes is down.
git-svn-id: https://ncisvn.nci.nih.gov/svn/oce_wcmteam/ContentDeliveryEngine/branches/Griffin@14539 e9cc8610-1bbf-714c-a5f3-26df0ede829a

+++ CDEFramework/Libraries/NCILibrary/Code/NCILibrary.Search.ElasticSearch/ESSiteWideSearchResultCollection.cs
@@ -11,16 +11,10 @@ namespace NCI.Search
     {
         private IEnumerable<ESSiteWideSearchResult> searchResults;
 
-        public int ResultCount { get; set; }
+        public long ResultCount { get; set; }
 
         List<ESSiteWideSearchResult> SearchResults { get; set; }
-
-        //protected ESSearchResultCollection(IEnumerable<ESSearchResult> results)
-        //{
-        //    this.searchResults = results;
-
-        //}
-
+        
         public List<ESSiteWideSearchResult> ESSearchCollection()
         {
             SearchResults = (List<ESSiteWideSearchResult>)searchResults;
@@ -47,18 +41,7 @@ namespace NCI.Search
         }
 
         #endregion
-
-        //#region IEnumerable Members
-        ///// <summary>
-        ///// Get IEnumerable
-        ///// </summary>
-        ///// <returns>the Enumerable list</returns>
-        //IEnumerator IEnumerable.GetEnumerator()
-        //{
-        //    return searchResults.GetEnumerator();
-        //}
-
-        //#endregion
+               
 
         IEnumerator IEnumerable.GetEnumerator()
         {
